+++ packages/eyes-sdk-core/lib/new/close-all-eyes.js
@@ -1,12 +1,24 @@
+const DiffsFoundError = require('../errors/DiffsFoundError')
+const NewTestError = require('../errors/NewTestError')
+const TestFailedError = require('../errors/TestFailedError')
+const TestResults = require('../TestResults')
+
 function makeCloseAllEyes({runner}) {
-  return async function getResults() {
-    const results = await runner.getAllTestResults(false)
-    return results.getAllResults().map(results => {
-      const container = results.toJSON()
-      if (container.exception && container.exception.getTestResults && container.exception.getTestResults()) {
-        return container.exception.getTestResults().toJSON()
+  return async function closeAllEyes({throwErr = false} = {}) {
+    const results = await runner.getAllTestResults()
+    return results.map(result => {
+      result = result instanceof TestResults ? result.toJSON() : null
+
+      if (throwErr) {
+        if (result.status === 'Unresolved') {
+          if (result.isNew) throw new NewTestError(result)
+          else throw new DiffsFoundError(result)
+        } else if (result.status === 'Failed') {
+          throw new TestFailedError(result)
+        }
       }
-      return container.testResults
+
+      return result
     })
   }
 }