+++ packages/eyes-sdk-core/lib/new/close.js
@@ -1,7 +1,28 @@
+const DiffsFoundError = require('../errors/DiffsFoundError')
+const NewTestError = require('../errors/NewTestError')
+const TestFailedError = require('../errors/TestFailedError')
+
 function makeClose({eyes}) {
-  return async function close() {
-    const result = await eyes.close(false)
-    return result ? result.toJSON() : null
+  return async function close({throwErr = false} = {}) {
+    let results = await eyes.close(false)
+
+    if (!results) return []
+    else if (!Array.isArray(results)) results = [results]
+
+    return results.map(result => {
+      result = result.toJSON()
+
+      if (throwErr) {
+        if (result.status === 'Unresolved') {
+          if (result.isNew) throw new NewTestError(result)
+          else throw new DiffsFoundError(result)
+        } else if (result.status === 'Failed') {
+          throw new TestFailedError(result)
+        }
+      }
+
+      return results
+    })
   }
 }